requirements.txt
Flask==2.3.3
pymongo==3.12.0
python-dotenv==0.19.0
Werkzeug==2.3.7
schedule==1.1.0
gunicorn==20.1.0